test: Incorporated Pin tests into a single file.
