+++ .gitignore
@@ -40,6 +40,7 @@ htmlcov
 
 # Directories
 docs/
+tests/schematic_output/
 **/\.ipynb_checkpoints/
 
 # Mr Developer
